Fix compiler warnings for test int_matmult.
Change type of variable to handle hosts with 16-bit ints.

ChangeLog:

	* src/int_matmult/matmult.c (matrix): Change from int to long.

+++ src/int_matmult/matmult.c
@@ -27,7 +27,7 @@
 
 #define UPPERLIMIT 20
 
-typedef int matrix [UPPERLIMIT][UPPERLIMIT];
+typedef long matrix [UPPERLIMIT][UPPERLIMIT];
 
 int Seed;
 matrix ArrayA, ArrayB, ResultArray;
